:broom:feat: melhorar visualizacao de dados e relacionamentos
